Your program should read two lines from the console, a term, and a definition, that represent a card.
After that, the user inputs a line as an answer (a definition of the term on the card). Compare the user's answer with the correct definition and print the result.

The output of the program must contain one of two words:

wrong if the answer doesn't match the definition;
right if the answer matches the definition.